# [IR][RFC] Enabling swizzling for dot encoding in some specific case