@wenqiny 2025-07-07
Thank you for all your comments!

I understand there are two alternatives solutions:

1. Try to get the suitable `dot encoding` in `swizzleDotOperandLike`, I saw this method when I was working on this PR, I though it's hard for triton to find the correct `dot encoding` in it because we got a weird `RankedTensorType type` (it's a reshaped and transposed type) from its arguments, if we want to  get the suitable `dot encoding` in it, we may have to call `getPotentialDotType` method in the below code snippet, which was introduced in this PR.
https://github.com/triton-lang/triton/blob/7a4cfe7d8491ccc5d8fd3f6441ff0a5059630f84/lib/Dialect/TritonGPU/Transforms/Utility.cpp#L1095-L1097
2. Try to elide the `reshape` op, because it just removing an empty dimension (`size = 1`), but I don't know we should do it in which pass.

If you though this issue is valuable to fix, could you please tell me which solution is better (I guess the 2nd solution is more generic, because it will not only benefit this case, but also make the IR more clear in other scenarios), I would like to have a try, any suggestions and insights would be appreciated!