@manman-ren 2025-07-07
Yeah I hit performance issues with 3D tensors where the leading dimension is 1 and for my case, the kernel was emitted from Helion. I worked around it by modifying the source generated from Helion.

From earlier discussions with Thomas, "we need to change tensor shapes by backward propagating all uses of the loads" and one implementation choice is to write a pass that converts 3D to 2D tensors if leading dimension is 1 for dots.